Fix comments after review: improved structure of test files. README file updated accordingly.
